DEL new module and move logic to ncf_manager
