+++ ncf_manager/models/__init__.py
@@ -28,3 +28,4 @@ from . import res_currency
 from . import account_invoice
 from . import res
 from . import ir_sequence
+from . import ncf_registry
