+++ ncf_manager/security/ir.model.access.csv
@@ -6,3 +6,5 @@ account.access_account_journal_user,account.journal,account.model_account_journa
 account.access_account_journal_invoice,'account.journal invoice',account.model_account_journal,account.group_account_invoice,1,1,,
 base.access_ir_sequence_group_user,'employee ir_sequence',model_ir_sequence,base.group_user,1,1,,
 base.access_ir_sequence_date_range_group_user,'ir_sequence_date_range group_user',model_ir_sequence_date_range,base.group_user,1,1,,
+access_account_ncf_registry,access_account_ncf_registry,model_account_ncf_registry,account.group_account_invoice,1,1,0,0
+access_account_ncf_registry,access_account_ncf_registry,model_account_ncf_registry,account.group_account_manager,1,1,1,1
