+++ ncf_manager/views/ncf_registry_view.xml
